+++ packages/react/src/components/UpsellingKit/UpsellRequestResponseDialog/index.tsx
@@ -17,26 +17,30 @@ interface UpsellRequestResponseDialogProps {
   open: boolean
   onClose?: () => void
   success: boolean
-  errorMessage: {
-    title: string
-    description: string
-  }
-  successMessage: {
-    title: string
-    description: string
-    buttonLabel: string
-    buttonOnClick: () => void
-  }
+  errorMessage: ErrorMessageProps
+  successMessage: SuccessMessageProps
   nextSteps: NextStepsProps
   closeLabel: string
 }
 
-interface StepItemProps {
+export interface ErrorMessageProps {
+  title: string
+  description: string
+}
+
+export interface SuccessMessageProps {
+  title: string
+  description: string
+  buttonLabel: string
+  buttonOnClick: () => void
+}
+
+export interface StepItemProps {
   text: string
   isCompleted?: boolean
 }
 
-interface NextStepsProps {
+export interface NextStepsProps {
   title: string
   items: StepItemProps[]
 }