+++ packages/react/src/experimental/Forms/EntitySelect/index.stories.tsx
@@ -11,6 +11,7 @@ import {
 } from "./groups-avatar-name.factory"
 import { EntitySelect } from "./index"
 import {
+  EntityId,
   EntitySelectEntity,
   EntitySelectNamedGroup,
   EntitySelectProps,
@@ -92,7 +93,7 @@ export const Default = {
   args: defaultArgs,
   render: (props: ComponentProps<typeof EntitySelect>) => {
     const [loading, setLoading] = useState<boolean>(props.loading ?? true)
-    const [expandedElements, setExpandedElements] = useState<number[]>([])
+    const [expandedElements, setExpandedElements] = useState<EntityId[]>([])
     const [selected, setSelected] = useState<EntitySelectEntity[]>([
       {
         ...famousEmployees[0],
@@ -103,7 +104,7 @@ export const Default = {
       props.selectedGroup ?? "all"
     )
 
-    const onItemExpandedChange = (id: number, expanded: boolean) => {
+    const onItemExpandedChange = (id: EntityId, expanded: boolean) => {
       if (expanded) {
         setExpandedElements([id].concat(expandedElements))
       } else {
@@ -217,7 +218,7 @@ export const WithSelectedGroup = {
   } as EntitySelectProps,
   render: (props: ComponentProps<typeof EntitySelect>) => {
     const [loading, setLoading] = useState<boolean>(props.loading ?? true)
-    const [expandedElements, setExpandedElements] = useState<number[]>([])
+    const [expandedElements, setExpandedElements] = useState<EntityId[]>([])
     const [selected, setSelected] = useState<EntitySelectEntity[]>([
       {
         ...famousEmployees[0],
@@ -228,7 +229,7 @@ export const WithSelectedGroup = {
       props.selectedGroup ?? "all"
     )
 
-    const onItemExpandedChange = (id: number, expanded: boolean) => {
+    const onItemExpandedChange = (id: EntityId, expanded: boolean) => {
       if (expanded) {
         setExpandedElements([id].concat(expandedElements))
       } else {
@@ -275,12 +276,12 @@ export const SingleSelector = {
   render: (props: ComponentProps<typeof EntitySelect>) => {
     const [loading, setLoading] = useState<boolean>(props.loading ?? true)
     const [selected, setSelected] = useState<EntitySelectEntity | undefined>()
-    const [expandedElements, setExpandedElements] = useState<number[]>([])
+    const [expandedElements, setExpandedElements] = useState<EntityId[]>([])
     const [selectedGroup, setSelectedGroup] = useState<string>(
       props.selectedGroup ?? "all"
     )
 
-    const onItemExpandedChange = (id: number, expanded: boolean) => {
+    const onItemExpandedChange = (id: EntityId, expanded: boolean) => {
       if (expanded) {
         setExpandedElements([id].concat(expandedElements))
       } else {
@@ -335,7 +336,7 @@ export const AlwaysOpen = {
   } as EntitySelectProps,
   render: (props: ComponentProps<typeof EntitySelect>) => {
     const [loading, setLoading] = useState<boolean>(props.loading ?? true)
-    const [expandedElements, setExpandedElements] = useState<number[]>([])
+    const [expandedElements, setExpandedElements] = useState<EntityId[]>([])
     const [selected, setSelected] = useState<EntitySelectEntity[]>([
       {
         ...famousEmployees[0],
@@ -346,7 +347,7 @@ export const AlwaysOpen = {
       props.selectedGroup ?? "all"
     )
 
-    const onItemExpandedChange = (id: number, expanded: boolean) => {
+    const onItemExpandedChange = (id: EntityId, expanded: boolean) => {
       if (expanded) {
         setExpandedElements([id].concat(expandedElements))
       } else {
@@ -396,7 +397,7 @@ export const AlwaysOpenInForm = {
     alwaysOpen: true,
   } as EntitySelectProps,
   render: (props: ComponentProps<typeof EntitySelect>) => {
-    const [expandedElements, setExpandedElements] = useState<number[]>([])
+    const [expandedElements, setExpandedElements] = useState<EntityId[]>([])
     const [selected, setSelected] = useState<EntitySelectEntity[]>([
       {
         ...famousEmployees[0],
@@ -407,7 +408,7 @@ export const AlwaysOpenInForm = {
       props.selectedGroup ?? "all"
     )
 
-    const onItemExpandedChange = (id: number, expanded: boolean) => {
+    const onItemExpandedChange = (id: EntityId, expanded: boolean) => {
       if (expanded) {
         setExpandedElements([id].concat(expandedElements))
       } else {
@@ -450,7 +451,7 @@ export const WithCustomTrigger = {
   } as EntitySelectProps,
   render: (props: ComponentProps<typeof EntitySelect>) => {
     const [loading, setLoading] = useState<boolean>(props.loading ?? true)
-    const [expandedElements, setExpandedElements] = useState<number[]>([])
+    const [expandedElements, setExpandedElements] = useState<EntityId[]>([])
     const [selectedGroup, setSelectedGroup] = useState<string>(
       props.selectedGroup ?? "all"
     )
@@ -463,7 +464,7 @@ export const WithCustomTrigger = {
     const [numSelected, setNumSelected] = useState<number>(2)
     const [open, setOpen] = useState<boolean>(false)
 
-    const onItemExpandedChange = (id: number, expanded: boolean) => {
+    const onItemExpandedChange = (id: EntityId, expanded: boolean) => {
       if (expanded) {
         setExpandedElements([id].concat(expandedElements))
       } else {
@@ -611,7 +612,7 @@ export const HiddenAvatar = {
     hiddenAvatar: true,
   },
   render: (props: ComponentProps<typeof EntitySelect>) => {
-    const [expandedElements, setExpandedElements] = useState<number[]>([])
+    const [expandedElements, setExpandedElements] = useState<EntityId[]>([])
     const [selected, setSelected] = useState<EntitySelectEntity[]>([
       {
         ...famousEmployees[0],
@@ -622,7 +623,7 @@ export const HiddenAvatar = {
       props.selectedGroup ?? "all"
     )
 
-    const onItemExpandedChange = (id: number, expanded: boolean) => {
+    const onItemExpandedChange = (id: EntityId, expanded: boolean) => {
       if (expanded) {
         setExpandedElements([id].concat(expandedElements))
       } else {